Guitar branch

+++ src/main/java/cn/didano/base/model/Tb_student_inf.java
@@ -3,7 +3,7 @@ package cn.didano.base.model;
 
 public class Tb_student_inf {
 
-	private Integer height;
+	private Double height;
     private Integer weight;
     private String orgImgUrl;
     private String created;
@@ -11,19 +11,23 @@ public class Tb_student_inf {
 	public Tb_student_inf() {
 	
 	}
-	public Tb_student_inf(Integer height, Integer weight, String orgImgUrl, String created) {
+
+	public Tb_student_inf(Double height, Integer weight, String orgImgUrl, String created) {
 		super();
 		this.height = height;
 		this.weight = weight;
 		this.orgImgUrl = orgImgUrl;
 		this.created = created;
 	}
-	public Integer getHeight() {
+
+	public Double getHeight() {
 		return height;
 	}
-	public void setHeight(Integer height) {
+
+	public void setHeight(Double height) {
 		this.height = height;
 	}
+
 	public Integer getWeight() {
 		return weight;
 	}
